refactor(e2e): reorganize e2e tests
Split up e2e tests into several files, all referring to common
utilities. This will keep the structure more manageable as we add more
tests.
